packages/agent/requirements.txt
requests>=2.31.0
psutil>=5.9.8
pywin32>=306
PyYAML>=6.0.1
Pillow>=10.3.0
